moved compute_coi_weights to xayn-ai-coi
